Fix various compilation issues. Many, but not all, of the changes are things such as: * Import issues - remove unused imports, import order, etc. * Use CassandraRelevantProperties * Remove duplicate methods and unused methods with compilation errors * Adjust method, variable visibility * Use Clock.nanoTime() instead of System.nanoTime() * nowInSec is now a long instead of int * Use ByteComparable.Version.OSS50 in place of OSS41 * Use InMemoryTrie in place of MemtableTrie * Use o.a.c.io.util.File in place of java.io.File * Use TimeUUID instead of UUID

+++ src/java/org/apache/cassandra/index/IndexRegistry.java
@@ -228,7 +228,7 @@ public interface IndexRegistry
 
             @Nullable
             @Override
-            public Index.Indexer indexerFor(Predicate<Index> indexSelector, DecoratedKey key, RegularAndStaticColumns columns, int nowInSec, WriteContext ctx, IndexTransaction.Type transactionType, Memtable memtable)
+            public Index.Indexer indexerFor(Predicate<Index> indexSelector, DecoratedKey key, RegularAndStaticColumns columns, long nowInSec, WriteContext ctx, IndexTransaction.Type transactionType, Memtable memtable)
             {
                 return null;
             }
